Misc type fixes

+++ api/deals/models.py
@@ -39,10 +39,10 @@ class DealModel(BaseModel):
     margin_short_loan_interest: float = Field(default=0)
     margin_short_buy_back_price: float = Field(default=0)
     margin_short_sell_qty: float = Field(default=0)
-    margin_short_buy_back_timestamp: int = 0
+    margin_short_buy_back_timestamp: float = 0
     margin_short_base_order: float = Field(default=0)
-    margin_short_sell_timestamp: int = Field(default=0)
-    margin_short_loan_timestamp: int = Field(default=0)
+    margin_short_sell_timestamp: float = Field(default=0)
+    margin_short_loan_timestamp: float = Field(default=0)
 
     # Refactored deal prices that combine both margin and spot
     opening_price: Optional[float] = Field(
@@ -53,7 +53,7 @@ class DealModel(BaseModel):
         default=0,
         description="replaces previous buy_total_qty or short_sell_qty/margin_short_sell_qty",
     )
-    opening_timestamp: Optional[int] = Field(default=0)
+    opening_timestamp: Optional[float] = Field(default=0)
     closing_price: Optional[float] = Field(
         default=0,
         description="replaces previous sell_price or short_sell_price/margin_short_sell_price",
@@ -62,7 +62,7 @@ class DealModel(BaseModel):
         default=0,
         description="replaces previous sell_qty or short_sell_qty/margin_short_sell_qty",
     )
-    closing_timestamp: Optional[int] = Field(
+    closing_timestamp: Optional[float] = Field(
         default=0,
         description="replaces previous buy_timestamp or margin/short_sell timestamps",
     )
